+++ components/modals/OneButtonModals.tsx
@@ -1,6 +1,56 @@
+import Icons from "@/constants/icons";
 import { useModal } from "@/hooks/useModal";
+import type { EmojiType } from "@/types/Modal.interface";
 import { useRouter } from "expo-router";
-import { OneButtonModal } from "./ModalProvider";
+import { TouchableOpacity } from "react-native";
+import { Text } from "react-native";
+import { View } from "react-native";
+
+const Emojis = {
+  SAD: <Icons.FaceNotDoneIcon width={40} height={40} />,
+  HAPPY: <Icons.FaceDoneIcon width={40} height={40} />,
+};
+
+export function OneButtonModal({
+  onClose,
+  emoji,
+  contents,
+  buttonText,
+  onPress,
+}: {
+  onClose: () => void;
+  emoji?: EmojiType;
+  contents: string;
+  buttonText: string;
+  onPress: () => void;
+}) {
+  return (
+    <View
+      className="h-full items-center justify-center px-7"
+      onTouchStart={onClose}
+    >
+      <View
+        className="items-center rounded-xl bg-white px-7 py-6"
+        onTouchStart={(e) => e.stopPropagation()}
+      >
+        {!!emoji && Emojis[emoji]}
+
+        <Text className="title-3 mt-4 text-center text-gray-90">
+          {contents}
+        </Text>
+
+        <TouchableOpacity
+          onPress={onPress}
+          className="mt-5 h-[52px] w-full grow flex-row items-center justify-center rounded-[8px] bg-primary"
+        >
+          <Text className="text-center font-pbold text-[17px] text-white leading-[150%]">
+            {buttonText}
+          </Text>
+        </TouchableOpacity>
+      </View>
+    </View>
+  );
+}
 
 export function EmailCheckModal() {
   const { closeModal } = useModal();
